实现删除用户userid前先查询

+++ src/main/java/com/example/login/service/impl/UserServiceImpl.java
@@ -66,8 +66,15 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public void deleteByUserId(String userId) {
+    public boolean deleteByUserId(String userId) {
+
+        User user = userRepository.findByUserId(userId);
+        if (user == null) {
+            return false;
+        }
+
         userRepository.deleteByUserId(userId);
+        return true;
     }
 
 }
